requirements.txt
asgiref==3.6.0
blinker==1.6.2
Brotli==1.0.9
click==8.1.3
colorama==0.4.6
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
Django==4.2.1
django-binary-database-files==1.0.10
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Flask==2.3.2
Flask-Compress==1.13
future==0.18.3
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
packaging==23.1
Pillow==9.5.0
plotly==5.14.1
psycopg2==2.9.6
PyJWT==2.7.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.2
tzdata==2023.3
Werkzeug==2.3.4
